New Codes :D

+++ src/XProjectProg6.java
@@ -5,18 +5,18 @@ public class XProjectProg6 {
         System.out.println("Enter a number");
         Scanner scanner=new Scanner(System.in);
         //Taking The Number as Input
-        int num=scanner.nextInt();
+        double num=scanner.nextDouble();
         System.out.println("Enter the exponent");
         //Taking The Exponent as Input
-        int exp=scanner.nextInt();
+        double exp=scanner.nextDouble();
         //Calling the 'power' method to calculate the number raised to the exponent
         double power=power(num, exp);
         //Printing the result
         System.out.println("Power = "+ power);
     }
     //Method 'power' for calculating number raised to the exponent
-    public static double power(int a, int b){
-        int i=1, x=1;
+    public static double power(double a, double b){
+        double i=1, x=1;
         //This if-block executes only for negative exponents
         if (b<0){
             b=b*(-1);
